refactor: rename & organize folder structure
