+++ src/components/PanelRight/PaneInfo.vue
